requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.1
automat==25.4.16
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
distro==1.9.0
django==5.2.8
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-prose-editor==0.22.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
faker==38.2.0
google-auth==2.43.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.11.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.3
langchain-core==1.2.0
langchain-openai==1.1.3
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
msgpack==1.1.2
nh3==0.3.2
openai==2.11.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
psycopg2-binary==2.9.10
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pyjwt==2.10.1
pyopenssl==25.3.0
python-dotenv==1.2.1
pyyaml==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
service-identity==24.2.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==14.0.1
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uuid-utils==0.12.0
uvicorn==0.38.0
websockets==15.0.1
xxhash==3.6.0
zope-interface==8.1.1
zstandard==0.25.0
